Changing how the filesystem is cleaned in opencast-clean
We leave the NFS mount alone here because we might not be managing it, and the storage might be on a remote system we
don't have direct access to.  Because of this we have to ensure that only a single node does the cleaning, otherwise we
see stale filehandles.  This also means that cleaning the filesystem happens in two passes - one on the local files and
the other on NFS.