Creating the storage directory wherever the Opencast user is being created.
There's a method to this madness: If we're creating the user than it's *very* likely that the NFS mount will be mounted
at some point.  This is important in cases where the NFS server is *not* an Opencast node itself, because then the
directory needs to have the correct UID:GID pair.